gui/requirements.txt
gradio
python-dotenv
tqdm
pandas
google-cloud-storage
numpy
